requirements.txt
discord.py>=2.0.0
yt-dlp>=2023.10.0
ffmpeg-python>=0.2.0
gTTS>=2.4.0
python-dotenv>=1.0.0
aiohttp>=3.9.0
PyNaCl>=1.5.0
